requirements.txt
Flask==3.1.0
kafka-python-ng==2.2.3
pyspark==3.5.5
python-dotenv==1.0.1
requests==2.32.3
scikit-learn==1.5.2
joblib==1.4.2
